Core/Networking: many updates from TC (#92)
* packet fixes to match tc 7.3.5

* lock free message buffer; other updates from tc master

+++ src/server/game/Server/Packets/ClientConfigPackets.cpp
@@ -20,7 +20,7 @@
 WorldPacket const* WorldPackets::ClientConfig::AccountDataTimes::Write()
 {
     _worldPacket << PlayerGuid;
-    _worldPacket << uint32(ServerTime);
+    _worldPacket << ServerTime;
     _worldPacket.append(AccountTimes, NUM_ACCOUNT_DATA_TYPES);
 
     return &_worldPacket;
